requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.46
psycopg[binary]==3.3.2
python-dotenv==1.0.1
requests==2.31.0
